better play and cancel button
