+++ src/feature/flow_chart_panel/components/CancelBtn.tsx
@@ -1,0 +1,22 @@
+import { Button } from "@src/components/ui/button";
+import useKeyboardShortcut from "@src/hooks/useKeyboardShortcut";
+import { Ban } from "lucide-react";
+
+type CancelBtnProps = {
+  cancelFC: () => void;
+};
+
+const CancelBtn = ({ cancelFC }: CancelBtnProps) => {
+  useKeyboardShortcut("ctrl", "c", cancelFC);
+  useKeyboardShortcut("meta", "c", cancelFC);
+
+  return (
+    <Button size="sm" variant="destructive" onClick={cancelFC}>
+      <Ban />
+      <div className="px-1"></div>
+      <div>Cancel</div>
+    </Button>
+  );
+};
+
+export default CancelBtn;
