--- src/feature/flow_chart_panel/components/play-btn/play-btn.css
@@ -1,71 +1,0 @@
-.btnPlay:before {
-  display: none;
-  top: -75%;
-  background-image: radial-gradient(circle, #00eaff 20%, transparent 20%),
-    radial-gradient(circle, transparent 20%, #00eaff 20%, transparent 30%),
-    radial-gradient(circle, #00eaff 20%, transparent 20%),
-    radial-gradient(circle, #00eaff 20%, transparent 20%),
-    radial-gradient(circle, transparent 10%, #00eaff 15%, transparent 20%),
-    radial-gradient(circle, #00eaff 20%, transparent 20%),
-    radial-gradient(circle, #00eaff 20%, transparent 20%),
-    radial-gradient(circle, #00eaff 20%, transparent 20%),
-    radial-gradient(circle, #00eaff 20%, transparent 20%);
-  background-size: 10% 10%, 20% 20%, 15% 15%, 20% 20%, 18% 18%, 10% 10%, 15% 15%,
-    10% 10%, 18% 18%;
-}
-.btnPlay:after {
-  display: none;
-  bottom: -75%;
-  background-image: radial-gradient(circle, #00eaff 20%, transparent 20%),
-    radial-gradient(circle, #00eaff 20%, transparent 20%),
-    radial-gradient(circle, transparent 10%, #00eaff 15%, transparent 20%),
-    radial-gradient(circle, #00eaff 20%, transparent 20%),
-    radial-gradient(circle, #00eaff 20%, transparent 20%),
-    radial-gradient(circle, #00eaff 20%, transparent 20%),
-    radial-gradient(circle, #00eaff 20%, transparent 20%);
-  background-size: 15% 15%, 20% 20%, 18% 18%, 20% 20%, 15% 15%, 10% 10%, 20% 20%;
-}
-.btnPlay:active {
-  transform: scale(0.8);
-  background-color: #00eaff;
-  box-shadow: 0 2px 25px rgba(5, 251, 242, 0.2);
-}
-.btnPlay.animate:before {
-  display: block;
-  animation: topBubbles ease-in-out 0.75s forwards;
-}
-.btnPlay.animate:after {
-  display: block;
-  animation: bottomBubbles ease-in-out 0.75s forwards;
-}
-
-@keyframes topBubbles {
-  0% {
-    background-position: 5% 90%, 10% 90%, 10% 90%, 15% 90%, 25% 90%, 25% 90%,
-      40% 90%, 55% 90%, 70% 90%;
-  }
-  50% {
-    background-position: 0% 80%, 0% 20%, 10% 40%, 20% 0%, 30% 30%, 22% 50%,
-      50% 50%, 65% 20%, 90% 30%;
-  }
-  100% {
-    background-position: 0% 70%, 0% 10%, 10% 30%, 20% -10%, 30% 20%, 22% 40%,
-      50% 40%, 65% 10%, 90% 20%;
-    background-size: 0% 0%, 0% 0%, 0% 0%, 0% 0%, 0% 0%, 0% 0%;
-  }
-}
-@keyframes bottomBubbles {
-  0% {
-    background-position: 10% -10%, 30% 10%, 55% -10%, 70% -10%, 85% -10%,
-      70% -10%, 70% 0%;
-  }
-  50% {
-    background-position: 0% 80%, 20% 80%, 45% 60%, 60% 100%, 75% 70%, 95% 60%,
-      105% 0%;
-  }
-  100% {
-    background-position: 0% 90%, 20% 90%, 45% 70%, 60% 110%, 75% 80%, 95% 70%,
-      110% 10%;
-    background-size: 0% 0%, 0% 0%, 0% 0%, 0% 0%, 0% 0%, 0% 0%;
-  }
-}
